requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.110.0
frozenlist==1.4.1
h11==0.14.0
idna==3.6
multidict==6.0.5
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
requests==2.31.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
upstash-redis==1.0.0
urllib3==2.2.1
uvicorn==0.28.1
yarl==1.9.4
